#25 [del] Delete FindIdRepository File etc
