+++ app/src/main/java/com/jjbaksa/jjbaksa/di/DataSourceModule.kt
@@ -4,9 +4,7 @@ import android.content.Context
 import com.jjbaksa.data.api.AuthApi
 import com.jjbaksa.data.api.NoAuthApi
 import com.jjbaksa.data.database.UserDao
-import com.jjbaksa.data.datasource.local.FindIdLocalDataSource
 import com.jjbaksa.data.datasource.local.UserLocalDataSource
-import com.jjbaksa.data.datasource.remote.FindIdRemoteDataSource
 import com.jjbaksa.data.datasource.remote.UserRemoteDataSource
 import dagger.Module
 import dagger.Provides
@@ -27,15 +25,5 @@ object DataSourceModule {
     @Singleton
     fun provideLocalDataSource(@ApplicationContext context: Context, userDao: UserDao): UserLocalDataSource {
         return UserLocalDataSource(context, userDao)
-    }
-    @Provides
-    @Singleton
-    fun provideFindIdRemoteDataSource(authApi: AuthApi, noAuthApi: NoAuthApi): FindIdRemoteDataSource {
-        return FindIdRemoteDataSource(authApi, noAuthApi)
-    }
-    @Provides
-    @Singleton
-    fun provideFindIdLocalDataSource(authApi: AuthApi, noAuthApi: NoAuthApi): FindIdLocalDataSource {
-        return FindIdLocalDataSource()
     }
 }
