+++ app/src/main/java/com/jjbaksa/jjbaksa/di/RepositoryModule.kt
@@ -1,13 +1,9 @@
 package com.jjbaksa.jjbaksa.di
 
-import com.jjbaksa.data.datasource.local.FindIdLocalDataSource
 import com.jjbaksa.data.datasource.local.UserLocalDataSource
-import com.jjbaksa.data.datasource.remote.FindIdRemoteDataSource
 import com.jjbaksa.data.datasource.remote.UserRemoteDataSource
-import com.jjbaksa.data.repository.FindIdRepositoryImpl
 import com.jjbaksa.domain.repository.UserRepository
 import com.jjbaksa.data.repository.UserRepositoryImpl
-import com.jjbaksa.domain.repository.FindIdRepository
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -24,13 +20,5 @@ object RepositoryModule {
         userLocalDataSource: UserLocalDataSource,
     ): UserRepository {
         return UserRepositoryImpl(userRemoteDataSource, userLocalDataSource)
-    }
-    @Singleton
-    @Provides
-    fun provideFindIdRepository(
-        findIdRemoteDataSource: FindIdRemoteDataSource,
-        findIdLocalDataSource: FindIdLocalDataSource
-    ): FindIdRepository {
-        return FindIdRepositoryImpl(findIdRemoteDataSource, findIdLocalDataSource)
     }
 }
