+++ data/src/main/java/com/jjbaksa/data/datasource/UserDataSource.kt
@@ -1,5 +1,6 @@
 package com.jjbaksa.data.datasource
 
+import com.jjbaksa.data.model.findid.FindIdResp
 import com.jjbaksa.data.model.user.LoginResp
 import com.jjbaksa.domain.resp.user.LoginReq
 import com.jjbaksa.domain.resp.user.SignUpReq
@@ -10,6 +11,8 @@ interface UserDataSource {
     suspend fun postSignUp(signUpReq: SignUpReq): Response<SignUpResp>?
     suspend fun checkAccountAvailable(account: String): Response<Unit>
     suspend fun postLogin(loginReq: LoginReq): Response<LoginResp>?
+    suspend fun checkAuthEmail(email: String): Response<Unit>
+    suspend fun findAccount(email: String, code: String): Response<FindIdResp>
     suspend fun saveAccessToken(accessToken: String)
     suspend fun saveAccount(account: String)
     suspend fun savePassword(password: String)
